monster: rework choose_nearby_injured_kin()
The old implementation of this function constructed a set of all nearby
eligible kin, then chose a member at random from that set. The new
implementation instead uses reservoir sampling[1] to avoid building an
intermediate data structure. Since this was the only use of z-set, this
change also deletes z-set in its entirety.

This change also adds a unit tests for the kin-choosing code, and in the
process of doing so adds two new test utility functions:

* t_build_arena(), which constructs an empty level with a wall around it
* t_add_monster(), which constructs a monster with a lot less setup than
  the production code usually does

[1]: https://en.wikipedia.org/wiki/Reservoir_sampling