Change terrain use during generation to better match Sil 1.3
Use SQUARE_WALL_INNER for '#' in vaults.  Resolves https://github.com/NickMcConnell/NarSil/issues/245 .
'%' is for quartz in vaults.
Prevent streamers from changing secret doors.  That's another cause for https://github.com/NickMcConnell/NarSil/issues/68 and for disconnecting the internals of vaults.
Change code comments to match what is done and to flag potential issues with the tunneling, both in NarSil and Sil 1.3.