Better match Sil 1.3's interface for firing and throwing
Resolves https://github.com/NickMcConnell/NarSil/issues/255 .  Resolves https://github.com/NickMcConnell/NarSil/issues/310 .

'f' is for firing from the first quiver in all keysets.  'F' is for firing from the second quiver in all keysets.  Refueling the light source has been moved to control-f in all keysets.  control-t is automatic throwing (throw the first item designed for throwing in the inventory at either the current target, if set and in range, or the nearest monster in range).  Tunneling with the hjkl option moves from control-t to 'T'.  Can only fire from the quivers and not from the pack or the floor.  target_okay() now check the range constraint if the current target is a monster.

Fire at nearest remains as before since it is compatible with the constraint of only firing from the quiver.