Allow the use command to equip an unequipped wearable item
Like Sil 1.3, using a torch or lantern will prefer to tap it as a fuel source, if there is a similar light source equipped, rather than equipping it.  Resolves Infinitum's report here, http://angband.oook.cz/forum/showpost.php?p=161397&postcount=2 , of not being able to use an item to equip it.