Have artifact count match up after reloading with smithed artifacts
Resolves https://github.com/NickMcConnell/NarSil/issues/415 .

Tolerate but warn if the number of artifacts in the save file is less than expected when reloaded.  As a bit of tolerance for modding, tolerate z_info->a_max being zero when a smithed artifact is created but skip using an artifact index of zero to agree with how artifact indices are used.