CMake: have unit test implementation only depend on cmake
With no dependence on external perl or sh scripts, allow building and running the unit tests when building the Windows front end.  Also use more memorable names for the targets associated with a single test suite ("unittest-effects-chain" builds the effects/chain suite; "run-unittest-effects-chain" runs (and, if necessary, builds) that suite.