Put stored corners for a vault just inside the vault's boundaries
That matches Sil 1.3 (see its generate.c:2798-2802,2890-2894,3033-3037), where the corners are for a simple room (just inside its bounding walls), and should make tunneling attempts more likely to suceed.  Change the solid_rock() tests for vaults to match Sil 1.3's behavior for interesting rooms and lesser vaults.

For greater vaults, Sil 1.3 uses (y1 - 5, x1 - 5, y2 + 5, x2 + 5) as the bounds for the solid_rock() test (see its generate.c:3022).  Do not use that here as Sil 1.3's behavior is not consistent with its own limits on where the vault is placed on the level and would lead to possible out-of-bounds accesses to the level's arrays.