Port over Vanilla Angband's UN_KTRL_CAP
Do not currently use it since do not have clashes for inscriptions or context menu shortcuts.  There are some overlaps when control is stripped (fuel a light source, ^f, with the fire commands, f and F; forge an item in the Angband-like keyset with Rogue-like navigation, ^d, and drop, d).  Those do not appear in the same context menus, and there's no useful object inscriptions targeting both of the commands involved in an overlap.