Port Vanilla Angband's changes to command queue string management
The lifetime of string argument for a command submitted to the queue is now managed by the queue.  Prevents a use after free if the last command is repeated immediately after inscribing an object.  Also change how the queue handles requests to repeat the last command.  Have it remember the possible target for repetition rather than searching through the queue for it.