Microchasm's tiles: add tiles for missing Sil 1.3 regular objects
Those are jewelled crowns, filthy rags, broken swords, deathblades, and sceptres.  Put them on a new row, 0x90, since the other row with enough empty space is currently all terrain tiles.  Leave the first tile in 0x90 empty for now so it could be used for the pile indicator.  Resolves a portion of the first part of https://github.com/NickMcConnell/NarSil/issues/608 .