Make weapon_glows() a bit more efficient
Return early if the weapon has no slays.  Add parameter to control the loop size for the line of sight check - some checks (player's light radius, grid data for a glowing object) only need to check the object's grid.  Break out of the loop for the line of sight check as soon as a grid in the line of sight is found.