Mark ordinary weapons and armour as known in the tutorial - 1.3
For kind recall, mark the item as known in the same circumstances as the knowledge menu does.  With https://github.com/NickMcConnell/NarSil/pull/872 also in place for the 1.3 branch, causes the tunneling bonus for shovels and mattocks to be shown in kind recall and allows the tunneling bonus for the tutorial's shovel to be seen without first wielding it.  That better emulates what Sil 1.3 does.