CMake: set maximum version with cmake_minimum_required() and cmake_policy() - 1.3
Avoids deprecation warning if using CMake 3.31.  Except for the scripts, use a maximum version of 3.26 since since src/cmake/modules/FindSphinx.cmake uses find_package(PythonInterp) and that needs to be changed with the policies introduced in 3.27.  For the scripts in src/cmake/scripts, use 3.31 as the maximum as there are no known incompatibilities.