CMake: add code coverage support - 1.3
Changes src/gen-coverage to use the arguments without modification.  That does not appear to affect "make coverage" with builds using configure and is necessary when that script is used with builds from CMake.

Add a bit of documentation in the compiling instructions for how to access the code coverage information on Linux/Unix.